+++ app/src/main/java/model/Maps.java
@@ -1,0 +1,5 @@
+package model;
+
+public abstract class Maps {
+    public abstract Enemy createEnemy();
+}
